+++ ask-sdk-model/ask_sdk_model/interfaces/audioplayer/player_activity.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 #
-# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file
 # except in compliance with the License. A copy of the License is located at
@@ -37,6 +37,7 @@ class PlayerActivity(Enum):
     BUFFER_UNDERRUN = "BUFFER_UNDERRUN"
     IDLE = "IDLE"
     STOPPED = "STOPPED"
+
     def to_dict(self):
         # type: () -> Dict[str, object]
         """Returns the model properties as a dict"""