+++ ask-sdk-model/ask_sdk_model/interfaces/monetization/v1/purchase_result.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 #
-# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file
 # except in compliance with the License. A copy of the License is located at
@@ -38,6 +38,7 @@ class PurchaseResult(Enum):
     NOT_ENTITLED = "NOT_ENTITLED"
     ERROR = "ERROR"
     ALREADY_PURCHASED = "ALREADY_PURCHASED"
+
     def to_dict(self):
         # type: () -> Dict[str, object]
         """Returns the model properties as a dict"""