+++ ask-sdk-model/ask_sdk_model/services/directive/directive_service_client.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 #
-# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file
 # except in compliance with the License. A copy of the License is located at
@@ -37,7 +37,6 @@ class DirectiveServiceClient(BaseServiceClient):
     :param api_configuration: Instance of :py:class:`ask_sdk_model.services.api_configuration.ApiConfiguration`
     :type api_configuration: ask_sdk_model.services.api_configuration.ApiConfiguration
     """
-
     def __init__(self, api_configuration):
         # type: (ApiConfiguration) -> None
         """
@@ -65,7 +64,9 @@ class DirectiveServiceClient(BaseServiceClient):
             raise ValueError(
                 "Missing the required parameter `send_directive_request` when calling `" + operation_name + "`")
 
-        resource_path = '/v1/directives'.replace('{format}', 'json')
+        resource_path = '/v1/directives'
+        resource_path = resource_path.replace('{format}', 'json')
+
         path_params = {}
 
         query_params = []