+++ ask-sdk-model/ask_sdk_model/services/service_client_factory.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 #
-# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file
 # except in compliance with the License. A copy of the License is located at
@@ -22,6 +22,7 @@ from .device_address import DeviceAddressServiceClient
 from .directive import DirectiveServiceClient
 from .list_management import ListManagementServiceClient
 from .monetization import MonetizationServiceClient
+from .proactive_events import ProactiveEventsServiceClient
 from .reminder_management import ReminderManagementServiceClient
 from .ups import UpsServiceClient